Session 121.3.2.6: Add Bulk Sync Endpoints (REST + CLI)
Implements GET /api/v1/sync/export and POST /api/v1/sync/import for bulk
data synchronization, plus CLI commands (pm sync export/import).

Architecture Changes:
- Move DTOs from pm-server to pm-core for better separation of concerns
- Add sync module to pm-core (ExportData, ImportResult, helper functions)
- Keep WorkItemRepository as executor-based (required for transaction support)

Backend:
- Add find_all() to all 7 repositories for bulk export
- Add Deserialize + TryFrom impls to all entity DTOs
- Implement export handler (7 bulk queries, no N+1)
- Implement import handler with conflict resolution (timestamp-based)
- Register routes at /api/v1/sync/export and /api/v1/sync/import

CLI:
- Add sync_commands module with Export and Import variants
- Add export_data() and import_data() client methods
- Add ClientError::Io variant with From<std::io::Error> for idiomatic error handling
- Wire command dispatch in main.rs

Testing:
- pm sync export --output data.json: exports all entities to JSON
- pm sync import --file data.json: idempotent import with conflict resolution

Docs:
- Update 121.3.2-Session-Plan.md: mark 121.3.2.6 as complete
- Update 121.3.2.6-Session-Plan.md: document actual implementation and learnings

+++ backend/crates/pm-ws/src/handlers/work_item.rs
@@ -112,7 +112,7 @@ pub async fn handle_create(
         item_type: item_type.clone(),
         parent_id,
         project_id,
-        position: max_position + 1,
+        position: (max_position + 1) as i32,
         title: sanitize_string(&req.title),
         description: req.description.as_ref().map(|d| sanitize_string(d)),
         status: req.status.clone().unwrap_or_else(|| "backlog".to_string()),
